requirements.txt
Django==1.11.5
djangorestframework==3.6.3
factory_boy==2.8.1
